# storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"VATUSA API","contact":{"name":"Daniel Hawton","url":"https:\/\/www.danielhawton.com"},"version":"2.0"},"host":"api.vatusa.net","basePath":"\/v2","schemes":["https"],"paths":{"\/auth":{"get":{"tags":["auth"],"description":"Get JWT.","produces":["application\/json"],"responses":{"200":{"description":"Return JSON Token.","schema":{"properties":{"token":{"type":"string","example":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJpc3MiOiJodHRwczovL2FwaS52YXR1c2EubmV0L3YyL2F1dGgiLCJpYXQiOjE1MTAyODA5NzIsImV4cCI6MTUxMDI4NDU3MiwibmJmIjoxNTEwMjgwOTcyLCJqdGkiOiIzRHU5S2xBRkJsQk5CTTA0Iiwic3ViIjoiODc2NTk0IiwicHJ2IjoiODdlMGFmMWVmOWZkMTU4MTJmZGVjOTcxNTNhMTRlMGIwNDc1NDZhYSJ9.vAzBMZgnzxymv6LnftovyN3Mww7obJ7Ms9H4nQ1la9dLOHpAzW2RvvBjMFdkvi3GyCJoVx23B8uOGOCpRKj5Qg"}},"type":"object"}},"401":{"description":"Unauthorized action.","$ref":"#\/definitions\/error"}}}},"\/exam\/queue\/{examId}":{"post":{"tags":["exam"],"summary":"Add exam to queue for the VATUSA Exam Center","description":"Sets the exam as the queued exam for VEC","produces":["application\/json"],"parameters":[{"name":"examId","in":"path","description":"Exam ID to place in queue","required":true,"type":"integer","format":"int64"},{"name":"bearer","in":"header","description":"JWT Token","required":true,"type":"string"}],"responses":{"404":{"description":"Exam assignment not found","schema":{"$ref":"#\/definitions\/error"},"examples":{"application\/json":{"status":"error","message":"Not Found"}}},"403":{"description":"Forbidden -- usually the exam assignment doesn't belong to the authenticated user","schema":{"$ref":"#\/definitions\/error"},"examples":{"application\/json":{"status":"error","message":"Forbidden"}}},"200":{"description":"Exam has been queued","schema":{"$ref":"#\/definitions\/OK"},"examples":{"application\/json":{"status":"OK"}}}},"security":["jwt"]}},"\/exam\/submit":{"post":{"tags":["exam"],"summary":"Submit exam payload for grading","description":"Submit exam from VEC for grading","produces":["application\/json"],"parameters":[{"name":"payload","in":"header","description":"Exam payload (base64)","required":true,"type":"string"},{"name":"answers","in":"header","description":"Answers (base64)","required":true,"type":"string"},{"name":"bearer","in":"header","description":"JWT Token","required":true,"type":"string"}],"responses":{"400":{"description":"Bad Request, usually for missing parameter","schema":{"$ref":"#\/definitions\/error"},"examples":[{"application\/json":{"status":"error","message":"Missing data"}},{"application\/json":{"status":"error","message":"Signature doesn't match payload"}}]},"404":{"description":"Exam assignment not found","schema":{"$ref":"#\/definitions\/error"},"examples":{"application\/json":{"status":"error","message":"Not Found"}}},"200":{"description":"Exam has been processed","schema":{"properties":{"results":{"type":"string"}},"type":"object"},"examples":[{"application\/json":{"results":"Not Passed"}},{"application\/json":{"results":"Passed"}}]}}}},"\/exam":{"get":{"tags":["exam"],"summary":"Generates and sends exam payload for VATUSA Exam Center based on queued exam for JWT auth'd user","description":"Generates and sends exam payload for VATUSA Exam Center based on queued exam for JWT auth'd user","produces":["application\/json"],"parameters":[{"name":"bearer","in":"header","description":"JWT Token","required":true,"type":"string"}],"responses":{"404":{"description":"Queue\/Exam Assignment not found","schema":{"$ref":"#\/definitions\/error"},"examples":[{"application\/json":{"status":"error","message":"No exam queued"}},{"application\/json":{"status":"error","message":"No matching exam assignment"}}]},"200":{"description":"Exam generated","schema":{"properties":{"payload":{"description":"base64 encoded quiz payload, with signature appended","type":"string"}},"type":"object"}}},"security":["jwt"]}}},"definitions":{"error":{"properties":{"status":{"type":"string","example":"error"},"message":{"type":"string","example":"not_logged_in"},"exception":{"type":"string"}},"type":"object"},"OK":{"properties":{"status":{"type":"string","example":"OK"}},"type":"object"}},"securityDefinitions":{"jwt":{"type":"basic","description":"JSON Web Token","name":"bearer","in":"header"},"session":{"type":"basic","description":"Session cookie","name":"cookie"},"apiKey":{"type":"apiKey","description":"Defined API Key","name":"api_key","in":"query"}}}